Post feature

+++ src/controllers/cms/postDetailDelete.controller.js
@@ -1,11 +1,17 @@
 const { deleteById } = require('../../models/Post.model');
+const { deleteCategoriesOnPosts } = require('../../models/CategoriesOnPosts.model');
 
 async function action(req, res) {
-  const resDeleteById = await deleteById(Number(req.params.id));
-  if (resDeleteById.status) {
-    res.send({
-      resDeleteById
-    });
+  console.log(Number(req.params.id), Number(req.params.category));
+
+  const resDeleteCategoriesOnPosts = await deleteCategoriesOnPosts(Number(req.params.id), Number(req.params.category));
+  if (resDeleteCategoriesOnPosts.status) {
+    const resDeleteById = await deleteById(Number(req.params.id));
+    if (resDeleteById.status) {
+      res.send({
+        resDeleteById
+      });
+    }
   }
 }
